Add a hack to get the Alcor Omea camera working
The Alcor Omea is actually a repackaged TIS DMK 51AU02.AS
camera, which as an unfortunate feature of reporting the wrong pixel
pattern (it says BGGR while in reality it is GBRG). This leads
to the debayerization going wrong, causing massive colour problems
in the image.

This adds a new pixel pattern named "HAX", which a) sets the GBRG
pattern, and b) omits the check whether the user-given pattern matches
the one reported by the camera hardware. It's a very ugly hack,
but works for now.